+++ modules/compatibility/pom.xml
@@ -48,6 +48,13 @@
 
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>ignite-commons</artifactId>
+            <version>${revision}</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>ignite-indexing</artifactId>
             <scope>test</scope>
         </dependency>
