+++ modules/compatibility/src/test/java/org/apache/ignite/compatibility/clients/JdbcThinCompatibilityTest.java
@@ -24,7 +24,7 @@ import java.sql.Statement;
 import org.apache.ignite.Ignite;
 import org.apache.ignite.cache.query.SqlFieldsQuery;
 import org.apache.ignite.internal.IgniteEx;
-import org.apache.ignite.internal.util.typedef.internal.U;
+import org.apache.ignite.internal.util.CommonUtils;
 import org.apache.ignite.lang.IgniteProductVersion;
 import org.junit.Assume;
 import org.junit.runner.RunWith;
@@ -53,10 +53,10 @@ public class JdbcThinCompatibilityTest extends AbstractClientCompatibilityTest {
     @Override protected void beforeTest() throws Exception {
         super.beforeTest();
 
-        int majorJavaVer = U.majorJavaVersion(U.jdkVersion());
+        int majorJavaVer = CommonUtils.majorJavaVersion(CommonUtils.jdkVersion());
 
         if (majorJavaVer > 11) {
-            Assume.assumeTrue("Skipped on jdk " + U.jdkVersion(),
+            Assume.assumeTrue("Skipped on jdk " + CommonUtils.jdkVersion(),
                 VER_2_12_0.compareTo(IgniteProductVersion.fromString(verFormatted)) < 0);
         }
     }
