+++ modules/compatibility/src/test/java/org/apache/ignite/compatibility/persistence/IndexTypesCompatibilityTest.java
@@ -39,7 +39,7 @@ import org.apache.ignite.cluster.ClusterState;
 import org.apache.ignite.compatibility.IgniteReleasedVersion;
 import org.apache.ignite.configuration.CacheConfiguration;
 import org.apache.ignite.internal.IgniteEx;
-import org.apache.ignite.internal.util.typedef.internal.U;
+import org.apache.ignite.internal.util.CommonUtils;
 import org.apache.ignite.lang.IgniteInClosure;
 import org.junit.Assume;
 import org.junit.Test;
@@ -110,10 +110,10 @@ public class IndexTypesCompatibilityTest extends IndexAbstractCompatibilityTest 
     /** */
     @Test
     public void testQueryOldIndex() throws Exception {
-        int majorJavaVer = U.majorJavaVersion(U.jdkVersion());
+        int majorJavaVer = CommonUtils.majorJavaVersion(CommonUtils.jdkVersion());
 
         if (majorJavaVer > 11) {
-            Assume.assumeTrue("Skipped on jdk " + U.jdkVersion(),
+            Assume.assumeTrue("Skipped on jdk " + CommonUtils.jdkVersion(),
                     VER_2_12_0.compareTo(IgniteReleasedVersion.fromString(igniteVer)) < 0);
         }
