+++ modules/control-utility/pom.xml
@@ -55,6 +55,13 @@
 
         <dependency>
             <groupId>${project.groupId}</groupId>
+            <artifactId>ignite-grid-unsafe</artifactId>
+            <version>${revision}</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>${project.groupId}</groupId>
             <artifactId>ignite-spring</artifactId>
         </dependency>
