+++ modules/core/src/test/java/org/apache/ignite/testframework/junits/multijvm/JavaVersionCommandParser.java
@@ -19,7 +19,7 @@ package org.apache.ignite.testframework.junits.multijvm;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import org.apache.ignite.internal.util.typedef.internal.U;
+import org.apache.ignite.internal.util.CommonUtils;
 
 /**
  * Parses output of 'java -version'.
@@ -43,6 +43,6 @@ class JavaVersionCommandParser {
 
         String fullJavaVer = matcher.group(2);
 
-        return U.majorJavaVersion(fullJavaVer);
+        return CommonUtils.majorJavaVersion(fullJavaVer);
     }
 }
