+++ modules/unsafe/src/main/java/org/apache/ignite/internal/util/FeatureChecker.java
